Converted over to using Tweeps on_status method
Converted the script over to use Tweepy’s on_status method, which
should remove the need to have that huge error exception handling.
Right now error handling is super basic and just prints to screen for
trouble shooting, we should move that into the logging module and call
it a day. Also created a requirements file